requirements.txt
dj-database-url>=0.4.1
django>=2.2
gunicorn>=19.4.5
psycopg2>=2.6.1
django-widget-tweaks>=1.4
djangorestframework
whitenoise>=4.1.2
requests>=2.19.1
pygments>=2.2
model_mommy